+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `--format` flag to select output format from "table", "json" or a custom Go template. Defaults to "table".
 - `endpoint group list|ls` command to print endpoint groups.
   - `--format` flag to select output format from "table", "json" or a custom Go template. Defaults to "table".
+- `endpoint inspect` command to print endpoint info.
+  - `--format` flag to select output format from "table", "json" or a custom Go template. Defaults to "table".
 - `help` command to print global help.
 - `login` command to authenticate against Portainer server.
   - `--print` flag to print the retrieved auth token.